+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # dev-server-utils
 
+## 0.0.4
+
+### Patch Changes
+
+- updated metadata to output filePath. bugfix for how selected variant paths are compared
+
 ## 0.0.2
 
 ### Patch Changes